+++ app/routes/deleteitem.py
@@ -2,9 +2,9 @@ from app import app, db
 from flask import redirect, url_for, request, session
 from app.forms import deleteItemForm
 from app.models.user import User
+from app.models.item import Item
 
-
-@app.route('/deleteitem', methods=['POST'])
+@app.route('/deleteitem/', methods=['POST'])
 def deleteItem():
     # Associated category specified by query parameter.
     category_id = request.args.get('id')
@@ -12,9 +12,12 @@ def deleteItem():
     deleteitem = deleteItemForm()
     if deleteitem.validate_on_submit():
         item = Item.query.filter_by(id=deleteitem.deleteID.data).one()
-        db.session.delete(item)
-        db.session.commit()
-        print "Item {} Deleted".format(item.name)
-        return redirect(url_for('showCategory', id=category_id))
+        if session['user_id'] == item.user_id:
+            db.session.delete(item)
+            db.session.commit()
+            print "Item {} Deleted".format(item.name)
+            return redirect(url_for('showCategory', id=category_id))
+        else:
+            return redirect(url_for('showCategory', id=category_id))
     else:
         return redirect(url_for('showCategory', id=category_id))
